Properly format hint output for subcommands
Subcommands for `rmt-cli repos custom` were not printed out correctly.
This seems to be a bug in Thor and could be worked around as in
`base.rb`. The `subcommand_prefix` method operates with the respective
class name, so in order to avoid further string manipulation we opted
for renaming the classes.