API Fix: service name
Notes:

In glue, the api appends release_stage (Alpha/Beta/...) to the product's
friendly_name in the serializer. The service serializer expects the
product's friendly name without the release_stage, but as the appended
version is saved in the db and used, we need to make a clean version of
friendly_name.