Don't create the license directory if the repo has no license
The `replace_directory(@temp_licenses_dir,...)` call would happen
whether there were licenses downloaded or not. Unfortunately, YaST
checks for the existence of that directorty to see if a product has a
license, but since RMT would create the directory empty, YaST then fails
to read the license file and failed. Now, we only create the license
directory if any licenses were actually downloaded. Solves bsc#1087707.

This commit also ensures that the temp directories created for metadata
and for licenses are always removed. Closes #118.