Use the recommended systemd config
Puma specifies a recommended way of starting it with systemd at:
https://github.com/puma/puma/blob/5fb2df9144803f0f80ec1d9fb19d20d433706fef/docs/systemd.md#service-configuration

Type=simple can be used, and then there's not need to --daemon the Rails
server.
Additionally, we set the RAILS_LOG_TO_STDOUT variable, so that Rails
outputs its log to stdout. This allows for the logs to be saved and
visible in journalctl.

The PrivateTmp option "sets up a new file system namespace for the
executed processes and mounts private /tmp and /var/tmp directories
inside it that is not shared by processes outside of the namespace.
(...) all temporary files created by a service in these directories will
be removed after the service is stopped."

Since RMT creates some temporary directories while mirroring, with this
option we isolate them and make sure they get deleted when RMT is
stopped.

Closes #128.