Skip repository check on the `upgrade` call
The upgrade call also works as the downgrade (rollback).
In a rollback, the system already has all RPMs installed, it just needs
to tell RMT which products it has. For that, it's not necessary that the
repos be mirrored.
In the upgrade call, YaST should check whether the products are
available, so it will not call this endpoint with products that are
unavailable anyway.