Fix products wrongly showing as enabled for mirroring
See Bugzilla bug #1102193 for reference.

To avoid confusion: When a repository is enabled or not, we talk about
zypper (RPM) repositories, which ship with a default, either `enabled`
or `disabled`. As an example, typically "debug symbols" are `disabled`
by default.

Before, if there were no `enabled` repositories for a product (which
happens for very old products), `mirror?` would return true even though
there was nothing mirrored. This kinda made sense since we accept a
__product__ as `enabled` when all the __repos__ that are `enabled` by
default are "enabled for mirroring", because that's usually everything
one needs (by definition, since they other repos on the clients are
`disabled` by default anyway).

Now it will first check if there are repositories that are `enabled` by
default. And if that is true, it then checks if these are all "enabled
for mirroring". If there are no repositories that are `enabled` by
default, it will check if there are repositories that are `disabled` by
default, and consider the product `enabled` if all those repositories
are "enabled for mirroring". If there are no repositories at all, it
will just return `false`, since a product without repositories is not
understood by RMT anyway.