Fix problem when a successor gets dropped
When upgrading from SLED 12 SP4 + SDK to SLED 15 SP1, QA discovered that
no migration was offered for SDK (which should get migrated to the Dev
Tools module).

The problem was when generating the migrations
(`generate(migration_kind: :offline)`), we would first calculate the
successors for the base and SDK, receiving SLED 15 SP1 + Dev Tools 15
SP1 (and nothing else).
But then, the call to #sort_migrations got rid of Dev Tools, because Dev
Tools sits on top of Desktop Apps, which sits on top of Basesystem, and
those modules were not part of the migration. The #sort_migrations
method then doesn't know where to place the Dev Tools, and drops it from
the migration.

Now, what we do is calculate the successors of the installed products,
again receiving only SLED 15 SP1 + Dev Tools 15 SP1.
But before sorting, we add to the migration all migration extras and
recommended modules, so the migration ends up with the entire list of
required products (SLED, Dev Tools, Basesystem and Desktop Apps).
Finally,  #sort_migrations is able to correctly place Dev Tools.

Fixes https://bugzilla.suse.com/show_bug.cgi?id=1118579