Fix license mirroring issue
When mirroring license files, RMT currently looks for a "product.license"
directory in the repo's parent directory (i.e. "../product.license").
While most, if not all, SUSE-built repositories follow that convention,
it's not followed by all products.
The directory is not always called "product.license".
In fact, the licenses are stored in repo_url + ".license",
and it just happens that the repo urls built by SUSE end in
"product", so the licenses are in the directory "product.license".

In the case of the IBM DLPAR Utils and IBM DLPAR SDK modules, the repo
urls end in "ppc64le", so the licenses are stored in a directory named
"ppc64le.license".