Rake test should fail on error
Currently, if rspec fails, we continue on and give an exit code of 0,
that shouldn't happen.

I split the code coverage into their respective code suites: core and
engine. Core will fail if we're missing 100% code coverage in normal RMT
and engine will fail if the engine files are not 100% tested.