Drop completion on singulars of the subcommands
Thor accepts incomplete commands, e.g. `product` for `products`, but also `pr`.
These are not the "official" subcommands. Before this commit, only the singular
cases of these words would also be completed, but the required code was ugly
and the solution inconsistent, because all the other shorthands wouldn't allow
completion.