Add repomd_parse gem
Currently, the project is using a local project library for parsing XML
repository files and it was suggested to use the gem repomd_parser,
which extracted most of the funcionalities of the local lib.

As a first step, I'm adding the gem to the project, and in the next
commits I'll try to replace the local lib.