Add fixtures for scenarios w/ mirror_src=false
Currently, the fixtures files for tests with the RMT::Mirror class
doesn't have data for a repository containing packages whose
architecture is different form "x86_64".

As there is a business rule for not mirroring packages with architecture
type "src" when the customer doesn't want it (by setting
`mirror_src=false`), the tests were fragile (removing the verification
for `mirror_src` doesn't break any tests in the class).

As a first step to replace the class RMT::Rpm::PrimaryXmlParser, I
created some fixtures files for packages with `arch=src` and added their
references to the ".xml.gz" repodata files.

I followed the steps below:

  1. Create some fake ".rpm" files for mocking packages w/ "arch=src";
  2. Editthe `.xml.gz` repodata files by decompressing them, adding the
     information about the new packages, replacing sha256 checksums for
     all the affected files and then compressing again the XML files to
     `.gz`;
  3. Remove the VCR cassette files and run a local server for providing
     the files (`cd spec/fixture/files` and the run the server with
     `python3 -m http.server 80`);
  4. Assure that every test pass green.

After all this, I removed the `mirror_src=false` check on the
RMT::Rpm::PrimaryXmlParser and some tests started to break. I assume
it is enmough to safely start replacing the parser by the gem.

P.S.: The fake packages data on the `primary.xml.gz` file and the new
packages themselves probably are very far from reflecting a real
situation, but for the purposa of testing the expected behavior, it
worked fine.