Replace RMT::Rpm::PrimaryXmlParser w/ gem
After adding the repomd_parser gem and replacing the project lib class
RMT::Rpm::RepomdXmlParser, I replaced the RMT::Rpm::PrimaryXmlParser
(used on RMT::Mirror) with an equivalent parser from the gem.

However, there is a subtle detail in the RMT::Rpm::PrimaryXmlParser
(also implemented in the RMT::Rpm::DeltainfoXmlParser): these parsers
doesn't return referenced files whose architecture is "src" if the
option `mirror_src` is set to false.

In the discussion about replacing the project lib parsers with the
repomd_parser, it was pointed that the `mirror_src` parameter only makes
sense in the context of this project (RMT), so it should not be included
in the gem.

To maintain the business rule (the customer has the option to not mirror
"arch=src" packages), I created a private method in the RMT::Mirror
class, `#filter_elegible_files` to filter elegible files for mirroring.
By now, the only condition for filtering the files is the combination of
arch=src and mirror_src=false, but in a future, the method could have
more conditions/rules for filtering the files.