Add test scenario for mirror_src=true
I added some fixtures files for packages with `arch=src` and assured the
business rule for not downloading these packages when `mirror_src=false`
was still working after replacing the project XML parsers with a gem.

However, there were no test scenarios for when `mirror_src=true`. I
added a scenario for this, and recorded again one of VCR cassettes to
allow requests for downloading RPM packages with `arch=src`. Now, if the
verification for `mirror_src=false` is removed, the tests will fail,
since more packages will be downloaded than the expected.