Rewrite some methods to improve readability
After I implmented the use of the repomd_parser gem in the RMT::Mirror
class, I took the opportunity to reorganize the methods which depend on
parsing XML files.

The goal of this commit was to improve the readability of these methods,
by breaking them into smaller methods or by simplifying some of the code
by using the Ruby core interface (mainly the Enumerable interface).

Specifically on the RMT::Mirror#mirror_data protected method, the same
logic was being applied to parse both primary and deltainfo files and
then download the packages, it was duplicated code. I split the method
into smaller methods, trying to tell the sequence in which the things
happen:

  1. Parse an XML file and get the name of all RPM packages [method
     `parse_mirror_data_files`];
  2. Filter what packages should be downloaded (considering `mirror_src`
     and possible duplicated files) [method `filter_eligible_packages`];
  3. Download the eligible-for-download packages [method
     `download_package_files`].