Add Ruby 2.6/2.7 to Travis CI pipeline
Currently, the build jobs for rmt-server on OBS are failing for openSUSE
Factory and Tumbleweed, because Ruby 2.5 packages are no longer
available for these distributions.

I added Ruby 2.6 and 2.8 to the CI pipeline to ensure the tests will not
fail for this Ruby version, before any modification on the OBS
configuration.