Increase gem 'typhoeus' version
The gem `typhoeus` uses `ethon` as a dependency. The version `0.11.0`
(Gemfile.lock) uses a method which is generating warnings in Ruby 2.7:

"rb_safe_level will be removed in Ruby 3.0"

I increased the gem `typhoeus` version to the latest release in order to
also increase `ethon` version, which fix the warning.