Refactor RMT::ChecksumVerifier class
The class RMT::Downloader relied on the RMT::ChecksumVerifier method
`.verify_checksum`, which raised an exception when the expected checksum
didn't match the actual one.

In order to reuse the existing checksum verification logic in different
classes, I moved the exception for unmatched checksums from the
ChecksumVerifier class to the Downloader one, since throwing it only
makes sense in the file downloading context. Also, I renamed the
`.verify_checksum` method to `.match_checksum?`, making it return a
boolean value.