Add logic to remove mismatched files/entries
During mirroring, even though RMT checks if there are entries that match
a given local path, it wouldn't remove duplicated entries or a
mismatched file.

Initially, the idea was to enforce uniqueness on the `local_path` field
of the `downloaded_files` table, but it could cause disruption on any
customer's database which has duplicated entries prior to it. Instead,
we opted to check for duplicated entries only when mirroring a file.

Then, I added some treatment for duplicated entries on the database and
for mismatched existing files on disk.