Add uniqueness constraint to `local_path`
As an improvement to downloaded files data integrity, I added a
uniqueness constraint to the field `local_path` on the  database table
`downloaded_files`.

Currently, the DownloadedFile model has a method to validate files which
removes duplicates "on the fly". I'm going to remove this logic it since
we enforced uniqueness on the database itself.