Rewrite DownloadedFile.add_file method
In order to make the code more semantic and to avoid duplicating
checking logic before adding new records to DownloadFile, I rewrote the
method `.add_file`, renaming it to `.track_file` and making it first
search for a record with a given 'local_path' and then update and
existing record or create a new one.

Additionally, I created the method `.untrack_file`, which is going to be
used in future new classes.