Create RMT file validator class
The DownloadedFile model had too much logic on file validness,
manipulating files on disks and checking their integrity (size,
checksum and presence).

In order to not bloat the model class with responsibilities which
doesn't concern the database itself, I created the class
::RMT::FileValidator which, right now, contains only one method to
validate and keep data integrity of local files both on disk and on
database.