Create #find_valid_file_by_checksum method
Currently, the class RMT::Deduplicator holds the logic for validating
a candidate file as a source file for deduplication. However, this logic
has some weak points:

* If a file with the given checksum is found on the database but is
  missing on the disk, the process ignores the deduplication but keep an
  invalid record on the database;
* It only validates a file by comparing the size on database with the
  file size on disk (does not allow verifying the checksum).

Since there's the new class RMT::FileValidator, which already has a more
elaborated logic for validating files on both disk and on database, I
create the method #find_valid_file_by_checksum, which will allow to both
find files for deduplication and to fix data integrity issues on the
fly.

Additionally, I create the method DownloadedFile#size as an alias to the
method #file_size, to allow polimorphism between RMT::FileReference and
DownloadedFile instances (the name 'file_size' in the context of a
DownloadedFile is redundant).