Improve execution time for local_path migration
The migration to add an unique index to `downloaded_files`.`local_path`
was taking about 90 seconds to complete and about 70 seconds were spent
in the duplicates query on a table with 300k+ records.

In the review, it was suggested to add a non-unique index before
querying for duplicates, and once the table is clean, remove the index
and readd it as an unique one.

It now takes about 23 seconds for each 'add index' operation plus 3
seconds for the duplicates query, which now deletes all records in one
single statement.