Fix thor deprecation warning
In the upgrade to thor 1.0.0, we missed a deprecation warning for thor.
If we don't define exit_on_failure? for RMT::CLI::Main, thor gives a
deprecation warning when it encounters a failure.

To keep previous behaviour, the recommendation is to return false.
However, as we were returning 0 for errors, the previous behaviour
wasn't correct. To fix that, let's return true instead.

See: https://github.com/erikhuda/thor/blob/master/CHANGELOG.md#100