Make rmt-cli exit w/ 1 on repo mirroring error
Currently, the command `rmt-cli mirror` doesn't exit with exit code 1
when it fails to mirror a repository; actually, it just logs a warn to
STDERR and continues execution until all repositories have been
mirrored. This can mess with RMT systemd service units or scripts, as
mentioned in issue #334.

I rewrote the error handling to summarize all error occurrences at the
end of the process instead of logging single messages in real time,
making the command exit with code 1.