Allow to Mirror Single Custom Repository
Before, it was only possible to mirror a single scc repository. The
problem was that scc id was used, so it was hard to single out a custom
repository. To allow for custom repos, we can just use the external URL,
as it's unique and possible to get from `rmt-cli repos custom ls`.