Remove RMT::CLI::Mirror#find_repo_by_target
The method RMT::CLI::Mirror#find_repo_by_target turned to be harder to
understand and to track potential errors. Also, it introduced a
potential bug where if it received `nil` as argument, it would return a
set of all custom repos, since their `scc_id` field are null.

I removed the method and took the logic to find a repo by its ID to
where it's used, the #repository method.