Make a singular id column for scc/custom repos
Before, we used two separate id columns for SCC and custom repositories.
We did this because it made SCC repoitories easier to use, as the
IDs would be the same across SCC/RMT/SMT. We didn't find a solution for
custom repositories, so we used the id field. This wasn't perfect, as
our repository handling diverged and it made it nearly impossible to
target both repo types in joint commands.

Now, we created a new column friendly_id which should make working with
repositories easier to use. For scc repositories, this column is filled
with the scc_id, for custom repositories, the MD5 of the url. This
should make both repository types easier to target, reduce the
ammount of copypasta, and ensure that the custom repo IDs will be the same
across RMT instances.

Note:

* scc_id was kept in the database to ensure backwards compatibility with
  previous versions of RMT, but it isn't used by the codebase. We should
  remove this leftover column in the future.