Backwards Compatibility: Old Custom Repo IDs
We've moved to a MD5 of the repository for the friendly_id, but users
still have the old repo IDs and may want to use them to target their
custom repositories.

This commit allows for users to target custom repositories with their
old IDs. For SCC repositories, users can only use the scc_id.