Protect Against Changing SCC Repo URLs
If SCC repositories changed their URLs, RMT had issues. This is because
we expect scc_id to be unique, and with a new URL, it tries to create a
new repository with the same scc_id. Instead of finding by the URL for
SCC repositories, we can find by scc_id instead and avoid this issue.