Issues with custom repos and UTF8 collation
The old regex method of finding matches didn't work because of the way
mysql collates utf8 content. Our regex would treat an "á" differently than
an "a", however, mysql treats them as the same. This could cause an
issue with duplicate entries with the same unique index. Instead, let's
have mysql compare the strings for us.