Raise an exception when there is an extension/module activated which has no migration successor configured.
This is the case for example for the LTSS extension. When a user with SLE12SP3 + LTSS calls `zypper migration`
he will receive an empty list of targets. With this patch, he will get shown an error message to de-activate
the LTSS extension first.

This copies commits 6369ef566cdb2ed72ee9b1005e5ab2c351c6a617 and
178018606b7ede4d186772040410f49d665c894c from glue