api: try to also pass the host login information to glue
If possible, try to pass the host's login information into a new header. This
header will be handled by glue so to link this registration proxy with the its
registered host.

If the host is not registered into SCC, nothing will happen and it will work as
usual. If the host is registered but the credentials file is not readable by the
running user, a warning will be printed, but everything will run as usual.

Signed-off-by: Miquel Sabaté Solà <msabate@suse.com>