Change expired subscription validation
Currently, the validation of expired subscription in the RMT API happens
through a `before_action` hook in the subscriptions/systems route.
However, the solution brought some undesirable side-effects which would
require some effort to keep the current API compatibility.

With this commit, a new approach is proposed, which checks the
subscription expiration onlye when creating an activation in RMT, the
very moment a subscription is actually used to be associated to a
system.