Fix `systems list` CSV output
The command `rmt-cli systems list --csv` had an issue with its output,
where the value in the column "Products" was presented as a list of
items separated by new line characters. This is counter-intuitive since
one expects a CSV file to have just one entry per line.

This commit fix the behavior by separating the items in the "Products"
columns by spaces.