Update the last_seen_at column on service refresh
Some systems might not use SUSEConnect ever again but still be active.
Before this commit, we only assumed that systems using SUSEConnect from
time to time were active, and we made sure about this with the optional
service that runs `SUSEConnect --keepalive`.

That being said, refreshing services from zypper also contact RMT, and
that's another point where `last_seen_at` can be updated as well.
However, this was not going through the existing authentication
workflow, and thus we could not figure out the system performing the
action. But, as it turns out, Zypper will actually be able to follow the
mechanism even if we add it now. This commit only adds this for Zypper,
since we know that it already works and we want to keep backwards
compatibility.

Signed-off-by: Miquel Sabaté Solà <msabate@suse.com>