Setup command for rmt-cli
The "bin/rmt-cli setup" feature adds a command to the RMT configuration
that allows users to interactively input specific values like database
and SCC credentials, and save them to the /etc/rmt.conf file. This makes
it easier for users to configure their RMT settings, especially for new
users or customers who may not be familiar with the configuration files.