Disable authentication for license files
We want to have a where the client, for example SUSEConnect, can display
license information to the user prior to registering a specific module or
product. With the intend to have this display prior to registration
the license file to be displayed cannot be in a path that requires
authentication. Update the nginx configuration for Public Cloud setup to
skip authentication for the license path.