Update cache scrubber
We will handle the system cache different for access on the repository and
registry path. File cache will be in locations determined by a file
written from the rails app. The app will also set the expiry time of the
cache. Update the scrubber to use the file written by RMT.