Allow sqlite3 deployments from scratch
Before this commit, if you tried to spin up a new RMT with sqlite3 as an
adapter you would be facing several issues. The first would be to
manually set the adapter on `config/rmt.yml`, whereas an environment
variable would suffice, as it's done for other database-related configuration.

Migrations were also quite mysql-centric, and they are mostly not
relevant since sqlite3 deployments are not even supported. Hence, we can
safely skip most of the data migration paths.

Signed-off-by: Miquel Sabaté Solà <msabate@suse.com>