fix: Rewrite update_proxy_byos migration
Rewrite the 'update_proxy_byos_column_type' migration to make it
compliant to 'strong_migrations' rules; adding a new column with a
non-null default value is considered a "dangerous operation."

The issue was caught when running 'docker-compose up' for local
development. The GitHub Actions pipeline doesn't run into this same
issue, and since the migration has been merged, not issues with it have
been reported; otherwise, a fresh-new installation of RMT would not be
properly set up.