feat: Adds configuration flag for .drpm mirroring
Add a configuration flag for enabling mirroring '.drpm' packages to the
RMT::Config class. The aimed behavior is that RMT doesn't mirror '.drpm'
packages by default, only when explictly required.

Currently, the flag isn't yet used in the code; it only "prepares the
ground" for the implementation of the desired behavior.