Fix missing public/repo path
With the removal of the public/repo/.gitignore file the public/repo
directory is no longer included in the git repo, causing make build
to fail due a missing public/repo directory.

To avoid problems cause by this issue, and additionally simplify the
use of the docker compose development environment, this patch adds
new make rules to ensure that both the public/repo path is created
and then public hierarchy is updated with appropriate permissions.
These new rules will be triggered as preconditions for the build
rule.

Update the DEVELOPMENT.md file to reflect these changes, and also
explicitly indicate that running the server rule leaves the server
running in the foreground of the user's shell session, and the same
is true for the shell rule.

Also update the toplevel .gitignore to explicitly ignore just the
public/repo directory, rather than any occurrence of a directory or
file with the name repo.