WIP: Add support for system data profiles
Add a new system_data_profiles table definition, with models defined
under app and spec.

Update v3 systems_controller to check for specific data profile types
in the hwinfo provided during a systems update, and if an entry is
found for a data profile type, ensure an corresponding entry exists
in the system data profiles table, creating a new entry if needed.
The hwinfo entry for the data profile type will be updated to remove
the profileData element, since the associated data is stored in the
system data profiles table.