Make data_profiles a parameter in the payload body
Modify the data profile handling to look for a data_profiles paramater
in the main params block of the request, rather than being a field in
the hwinfo parameter.

Add support for this in the system serializer.

After confirming that the appropriate system_data_profiles entry has
been created or already exists, update the hwinfo param to include an
entry for the associated data profile by appending "_profile" to the
<profile_type>