Avoid deadlock caused by racing registrations
If multiple registrations are attempting to add the same data profiles
we can sometimes see a deadlock as conflicting transactions are rolled
back and then retried.

To avoid this we first identify any data profiles that may need to be
created and attempt to create them all at one using as an upsert, to
ensure that all of them already exist before attempting to set up the
associations via the system_profiles table.

Additionally only retry a limited number of times before failing.

Tweaked last_seen_at updating to also use an upsert to ensure that
all are update at once.