Switch to response header indicating client submission issues
Ensure that request processing continues normally when provided data
profiles are invalid or incomplete by filtering out the problematic
data profiles before attempting to process them.

Additionally, if problematic data profiles are detected then add a
response header, X-System-Profile-Action, with the value cache-clear
that will indicate to the client that it needs to clear any cached
state associated with data profiles to ensure that it sends complete
profiles when it sends it's next request.

Add support for filtering a payload's data profiles into three
categories:
  * complete
  * incomplete
  * invalid

Add support for retrieving the existing data profiles associated
with a set of incomplete data profiles to assist in identifying
problematic data profiles that may have been provided in an update
request.

Fix rubocop warnings.