Address review feedback from Robert
Improve the debug messages logged by the announce_system and update
request handlers to report the profile types for problematic profiles
identified.

Additionally enhance the Profile.filter_profiles() method to return
hashes with symbolized keys to simplify determining which incomplete
profiles are unknown.

Only update the profiles associated with a system if valid complete
profiles were either provided or identified from incomplete profiles,
and add a test to ensure that existing profile associations are not
replaced if not valid complete profiles were provided in the update.