Enhance ensure_profile_exists() rescue handling
Under heavy stress test loading a subtle race was identified where
the find_by() in the rescue block was failing to find the Profile
record created by a racing request handler. By locking the table
before calling find_by() we ensure that the query is run against the
current state of the table, rather than a potentially stale read-only
version that was restored as part of the rollback, which wouldn't
include the yet to be created new record.