Avoid nested transactional overheads
The recent ensure_profile_exists() changes used nested transactions
to minimize the rollback cost of potential collisions, but doing so
incurs a general overhead that may not be beneficial unless racing
collisions occur very frequently. For now the nested transaction
wrapping has been converted to a comment about it's potential use
if the need arrises at a future point.