Add profiles support to scc_proxy hybrid announce_system
Enhance announce_base_product_hybrid() to include a system's profiles
in the announce_system request that it sends to the SCC.

Add a new as_payload() helper method to the Profile model that returns
the payload representation of a Profile record.

Update the SystemSerializer to leverage the as_payload() helper when
serializing systems.

Simplify a number of unit tests by leveraging the as_payload() helper
and a new test to verify correct operation of the helper.

Update scc_proxy unit tests to ensure that the inclusion of the system
profiles by announce_base_product_hybrid() is exercised.